Local fabio

+++ resources/views/users/datatables_actions.blade.php
@@ -1,15 +1,23 @@
-{!! Form::open(['route' => ['users.destroy', $id], 'method' => 'delete']) !!}
-<div class='btn-group'>
-    <a href="{{ route('users.show', $id) }}" class='btn btn-default btn-xs'>
-        <i class="far fa-eye"></i>
-    </a>
-    <a href="{{ route('users.edit', $id) }}" class='btn btn-default btn-xs'>
-        <i class="fas fa-edit"></i>
-    </a>
-    {!! Form::button('<i class="fas fa-trash"></i>', [
-        'type' => 'submit',
-        'class' => 'btn btn-danger btn-xs',
-        'onclick' => "return confirm('Tem certeza?')"
-    ]) !!}
-</div>
-{!! Form::close() !!}
+@if($name != 'Admin')
+    {!! Form::open(['route' => ['users.destroy', $id], 'method' => 'delete']) !!}
+    <div class='btn-group'>
+        <a href="{{ route('users.show', $id) }}" class='btn btn-default btn-xs'>
+            <i class="far fa-eye"></i>
+        </a>
+        <a href="{{ route('users.edit', $id) }}" class='btn btn-default btn-xs'>
+            <i class="fas fa-edit"></i>
+        </a>
+        {!! Form::button('<i class="fas fa-trash"></i>', [
+            'type' => 'submit',
+            'class' => 'btn btn-danger btn-xs',
+            'onclick' => "return confirm('Tem certeza?')"
+        ]) !!}
+    </div>
+    {!! Form::close() !!}
+@else
+    <div class='btn-group'>
+        <a href="{{ route('users.show', $id) }}" class='btn btn-default btn-xs'>
+            <i class="far fa-eye"></i>
+        </a>
+    </div>
+@endif
